+++ packages/@lou.codes/predicates/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change log
 
+## 1.0.8
+
+### Patch Changes
+
+-   🏗️ use argument destructing less to achieve smaller bundles.
+-   ⬆️ upgrade `@lou.codes/types` to `v1.1.1`
+-   ⬆️ upgrade `@lou.codes/parsers` to `v1.0.8`
+
 ## 1.0.7
 
 ### Patch Changes